+++ tree-sitter-haskell/TreeSitter/Haskell.hs
@@ -1,15 +1,13 @@
-{-# LANGUAGE TemplateHaskell #-}
 module TreeSitter.Haskell
 ( tree_sitter_haskell
-, Grammar(..)
+, getNodeTypesPath
 ) where
 
-import Language.Haskell.TH
-import TreeSitter.Haskell.Internal
+import Foreign.Ptr
 import TreeSitter.Language
+import Paths_tree_sitter_haskell
 
--- Regenerate template haskell code when these files change:
-addDependentFileRelative "../vendor/tree-sitter-haskell/src/parser.c"
+foreign import ccall unsafe "vendor/tree-sitter-haskell/src/parser.c tree_sitter_haskell" tree_sitter_haskell :: Ptr Language
 
--- | Statically-known rules corresponding to symbols in the grammar.
-mkSymbolDatatype (mkName "Grammar") tree_sitter_haskell
+getNodeTypesPath :: IO FilePath
+getNodeTypesPath = getDataFileName "vendor/tree-sitter-haskell/src/node-types.json"