+++ tree-sitter-haskell/internal/TreeSitter/Haskell/Internal.hs
@@ -1,8 +1,0 @@
-module TreeSitter.Haskell.Internal
-( tree_sitter_haskell
-) where
-
-import Foreign.Ptr
-import TreeSitter.Language
-
-foreign import ccall unsafe "vendor/tree-sitter-haskell/src/parser.c tree_sitter_haskell" tree_sitter_haskell :: Ptr Language