+++ tree-sitter-php/TreeSitter/PHP.hs
@@ -1,15 +1,13 @@
-{-# LANGUAGE TemplateHaskell #-}
 module TreeSitter.PHP
 ( tree_sitter_php
-, Grammar(..)
+, getNodeTypesPath
 ) where
 
-import Language.Haskell.TH
-import TreeSitter.PHP.Internal
+import Foreign.Ptr
 import TreeSitter.Language
+import Paths_tree_sitter_php
 
--- Regenerate template haskell code when these files change:
-addDependentFileRelative "../vendor/tree-sitter-php/src/parser.c"
+foreign import ccall unsafe "vendor/tree-sitter-php/src/parser.c tree_sitter_php" tree_sitter_php :: Ptr Language
 
--- | Statically-known rules corresponding to symbols in the grammar.
-mkSymbolDatatype (mkName "Grammar") tree_sitter_php
+getNodeTypesPath :: IO FilePath
+getNodeTypesPath = getDataFileName "vendor/tree-sitter-php/src/node-types.json"