+++ tree-sitter-php/internal/TreeSitter/PHP/Internal.hs
@@ -1,8 +1,0 @@
-module TreeSitter.PHP.Internal
-( tree_sitter_php
-) where
-
-import Foreign.Ptr
-import TreeSitter.Language
-
-foreign import ccall unsafe "vendor/tree-sitter-php/src/parser.c tree_sitter_php" tree_sitter_php :: Ptr Language