+++ tree-sitter-ruby/TreeSitter/Ruby.hs
@@ -1,15 +1,13 @@
-{-# LANGUAGE TemplateHaskell #-}
 module TreeSitter.Ruby
 ( tree_sitter_ruby
-, Grammar(..)
+, getNodeTypesPath
 ) where
 
-import Language.Haskell.TH
-import TreeSitter.Ruby.Internal
+import Foreign.Ptr
 import TreeSitter.Language
+import Paths_tree_sitter_ruby
 
--- Regenerate template haskell code when these files change:
-addDependentFileRelative "../vendor/tree-sitter-ruby/src/parser.c"
+foreign import ccall unsafe "vendor/tree-sitter-ruby/src/parser.c tree_sitter_ruby" tree_sitter_ruby :: Ptr Language
 
--- | Statically-known rules corresponding to symbols in the grammar.
-mkSymbolDatatype (mkName "Grammar") tree_sitter_ruby
+getNodeTypesPath :: IO FilePath
+getNodeTypesPath = getDataFileName "vendor/tree-sitter-ruby/src/node-types.json"