+++ tree-sitter-ruby/bench/Main.hs
@@ -1,7 +1,0 @@
-module Main (main) where
-
-import Gauge
-import qualified Parsing
-
-main :: IO ()
-main = defaultMain [ Parsing.benchmarks ]