+++ tree-sitter-typescript/TreeSitter/TypeScript.hs
@@ -1,15 +1,13 @@
-{-# LANGUAGE TemplateHaskell #-}
 module TreeSitter.TypeScript
 ( tree_sitter_typescript
-, Grammar(..)
+, getNodeTypesPath
 ) where
 
-import Language.Haskell.TH
-import TreeSitter.TypeScript.Internal
+import Foreign.Ptr
 import TreeSitter.Language
+import Paths_tree_sitter_typescript
 
--- Regenerate template haskell code when these files change:
-addDependentFileRelative "../vendor/tree-sitter-typescript/typescript/src/parser.c"
+foreign import ccall unsafe "vendor/tree-sitter-typescript/typescript/src/parser.c tree_sitter_typescript" tree_sitter_typescript :: Ptr Language
 
--- | Statically-known rules corresponding to symbols in the grammar.
-mkSymbolDatatype (mkName "Grammar") tree_sitter_typescript
+getNodeTypesPath :: IO FilePath
+getNodeTypesPath = getDataFileName "vendor/tree-sitter-typescript/typescript/src/node-types.json"