+++ tree-sitter-java/ChangeLog.md
@@ -1,3 +1,9 @@
+# v0.7.0.0
+
+* remove CodeGen and Template Haskell splice
+* move contents of internal library into library and remove internal library
+* add `getNodeTypesPath` to provide easier access to `node-types.json` file
+
 # v0.4.0.0
 
 * AST datatypes use a shared `Token` datatype for all anonymous leaves.